% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{depressiondata}
\alias{depressiondata}
\title{Global Depression Data}
\format{
A data frame with 6000 rows and 7 variables:
\describe{
\item{\code{Code}}{character Country Code}
\item{\code{Countries}}{character Country Name}
\item{\code{Years}}{integer Year of Observation}
\item{\code{Prevalence_depr_male}}{double Percentage of Males with Depression}
\item{\code{Prevalence_depr_female}}{double Percentage of Females with Depression}
\item{\code{Population_Estimate}}{double Estimate of Population}
\item{\code{Geometry}}{list For map plotting}
}
}
\source{
\url{http://raw.githubusercontent.com/lostconnectionhere/mental_health/main/data/prevalence-of-depression-males-vs-females.csv /}
}
\usage{
depressiondata
}
\description{
Dataset of Global Depression
}
\keyword{datasets}
