% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{labourdata}
\alias{labourdata}
\title{OECD Labour Data}
\format{
A data frame with 798 rows and 3 variables:
\describe{
\item{\code{Countries}}{character Country Name}
\item{\code{Years}}{double Year of Observation}
\item{\code{Hours}}{double Average Hours of Work Yearly}
}
}
\source{
\url{http://stats.oecd.org/index.aspx?DataSetCode=TUD/}
}
\usage{
labourdata
}
\description{
Dataset of OECD Countries and their average yearly work hours
}
\keyword{datasets}
