% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maindata}
\alias{maindata}
\title{Main Data}
\format{
A data frame with 63 rows and 9 variables:
\describe{
\item{\code{Countries}}{integer Country Name}
\item{\code{Years}}{double Year of Observation}
\item{\code{UnionDensity}}{double Percentage of Employees in Unions}
\item{\code{Wages}}{double Minimum Wage in USD}
\item{\code{Population_Estimate}}{double Estimate of Population}
\item{\code{Dep_Prevalence}}{double Percentage of Population with Depression}
\item{\code{Suicides_Per100k}}{double Number of Suicides per 100,000}
\item{\code{Hours}}{double Average Hours Worked per Year} 
}
}
\source{
\url{wrangled}
}
\usage{
maindata
}
\description{
Main dataframe containing all variables, use this for modelling
}
\keyword{datasets}
