% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minwagedata}
\alias{minwagedata}
\title{OECD Minimum Wage Data}
\format{
A data frame with 640 rows and 3 variables:
\describe{
\item{\code{Countries}}{character Country Name}
\item{\code{Years}}{character Year of Observation}
\item{\code{Wages}}{double Minimum Wage in USD}
}
}
\source{
\url{https://stats.oecd.org/index.aspx?DataSetCode=RMW}
}
\usage{
minwagedata
}
\description{
Dataset of OECD Countries and their minimum wage in USD
}
\keyword{datasets}
