% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{suicidedata}
\alias{suicidedata}
\title{Global Suicide Data}
\format{
A data frame with 6000 rows and 5 variables:
\describe{
\item{\code{Code}}{character Country Code}
\item{\code{Countries}}{character Country Name}
\item{\code{Years}}{integer Year of Observation}
\item{\code{Suicides_Per100k}}{double Number of Suicides per 100,000 people}
\item{\code{Geometry}}{list For map plotting}
}
}
\source{
\url{http://raw.githubusercontent.com/popsnot/DATA201-Group-Project-2022/main/suicide-death-rates.csv/}
}
\usage{
suicidedata
}
\description{
Dataset of Global Suicide Rates
}
\keyword{datasets}
