% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uniondata}
\alias{uniondata}
\title{OECD Union Density Data}
\format{
A data frame with 798 rows and 3 variables:
\describe{
\item{\code{Countries}}{character Country Name}
\item{\code{Years}}{double Year of Observation}
\item{\code{UnionDensity}}{double Percentage of Employees in Unions}
}
}
\source{
\url{http://stats.oecd.org/index.aspx?DataSetCode=TUD/}
}
\usage{
uniondata
}
\description{
Dataset of OECD Countries and the percentage of employees in unions
}
\keyword{datasets}
